Updated to support BehaveBio objects and executer
