+++ src/main/java/org/secureauth/sarestapi/data/Response/BehaveBioResponse.java
@@ -1,28 +1,61 @@
 package org.secureauth.sarestapi.data.Response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import org.secureauth.sarestapi.data.BehavioralBio.BehaveBioResults;
-import org.secureauth.sarestapi.data.Response.BaseResponse;
+import org.secureauth.sarestapi.data.BehavioralBio.BehaveBioResult;
 import org.secureauth.sarestapi.util.JSONUtil;
 
 import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlSeeAlso;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Created by rrowcliffe on 4/30/16.
  */
-@XmlRootElement
+@XmlRootElement (name="BehaviorBioResults")
+@XmlSeeAlso(BehaveBioResult.class)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class BehaveBioResponse extends BaseResponse{
 
-    private BehaveBioResults behaveBioResults;
+    private double TotalScore;
+    private double TotalConfidence;
+    private String Device;
+    private List<BehaveBioResult> Results = new ArrayList<BehaveBioResult>();
 
-
-    public BehaveBioResults getBehaveBioResults() {
-        return behaveBioResults;
+    public double getTotalScore() {
+        return TotalScore;
     }
 
-    public void setBehaveBioResults(BehaveBioResults behaveBioResults) {
-        this.behaveBioResults = behaveBioResults;
+    public void setTotalScore(double totalScore) {
+        TotalScore = totalScore;
+    }
+
+    public double getTotalConfidence() {
+        return TotalConfidence;
+    }
+
+    public void setTotalConfidence(double totalConfidence) {
+        TotalConfidence = totalConfidence;
+    }
+
+    public void setResults(List<BehaveBioResult> results) {
+        Results = results;
+    }
+
+    public String getDevice() {
+        return Device;
+    }
+
+    public void setDevice(String device) {
+        Device = device;
+    }
+
+    public List<BehaveBioResult> getResults() {
+        return Results;
+    }
+
+    public void setResults(ArrayList<BehaveBioResult> results) {
+        Results = results;
     }
 
     @Override
